Removed Complicated Variables
I removed some of the variables used to control the calibration. While it would be nice to have some of them, it makes it easier to test when the complexity is lessened. There may be some other issues to work out but I would recommend looking into the changes I've made in advance to lower the complexity and make it easier to troubleshoot.